feat:인증 상태 체크하기

+++ Chapter06/src/main/java/com/rubypaper/controller/BoardController.java
@@ -6,20 +6,23 @@ import org.apache.logging.log4j.message.ReusableMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Date;
+import com.rubypaper.domain.Member;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
     @Autowired
     private BoardService boardService;
     @RequestMapping("/getBoardList")
-    public String getBoardList(Model model, Board board){
+    public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         List<Board> boardList = boardService.getBoardList(board);
 
         model.addAttribute("boardList", boardList);
@@ -27,30 +30,45 @@ public class BoardController {
     }
 
     @GetMapping("/insertBoard")
-    public String insertBoardView(){
+    public String insertBoardView(@ModelAttribute("member") Member member){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         return "insertBoard";
     }
 
     @PostMapping("/insertBoard")
-    public String insertBoard(Board board){
+    public String insertBoard(@ModelAttribute("member") Member member, Board board){
+        if(member.getId() == null){
+            return "redirect:login";
+        }
         boardService.insertBoard(board);
         return "redirect:getBoardList";
     }
 
     @GetMapping("/getBoard")
-    public String getBoard(Board board, Model model){
+    public String getBoard(@ModelAttribute("member") Member member, Board board, Model model){
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         model.addAttribute("board", boardService.getBoard(board));
         return "getBoard";
     }
 
     @PostMapping("/updateBoard")
-    public String updateBoard(Board board){
+    public String updateBoard(@ModelAttribute("member") Member member, Board board){
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         boardService.updateBoard(board);
         return "forward:getBoardList";
     }
 
     @GetMapping("/deleteBoard")
-    public String deleteBoard(Board board){
+    public String deleteBoard(@ModelAttribute("member") Member member, Board board){
+        if (member.getId() == null) {
+            return "redirect:login";
+        }
         boardService.deleteBoard(board);
         return "forward:getBoardList";
     }
@@ -58,5 +76,10 @@ public class BoardController {
     @GetMapping("/hello")
     public void hello(Model model){
         model.addAttribute("greeting", "Hello Thymeleaf ^^!");
+    }
+
+    @ModelAttribute("member")
+    public Member setMember(){
+        return new Member();
     }
 }
